use existing init ramdisk

+++ usr/lib/archboot/iso.sh
@@ -57,9 +57,13 @@ _prepare_kernel_initramfs_files() {
     source "${_PRESET}"
     mkdir -p "${_ISODIR}"/EFI/{BOOT,TOOLS}
     mkdir -p "${_ISODIR}/boot"
-    archboot-cpio.sh -c "/etc/archboot/${_ARCH}-init.conf" -k "${ALL_kver}" -g "${_ISODIR}/boot/init-${_ARCH}.img" || exit 1
-    # save init ramdisk for further images
-    cp ${_ISODIR}/boot/init-${_ARCH}.img ./
+    if [[ -f "./init-${_ARCH}.img" ]]; then
+        cp ./init-${_ARCH}.img ${_ISODIR}/boot/
+    else
+        archboot-cpio.sh -c "/etc/archboot/${_ARCH}-init.conf" -k "${ALL_kver}" -g "${_ISODIR}/boot/init-${_ARCH}.img" || exit 1
+        # save init ramdisk for further images
+        cp ${_ISODIR}/boot/init-${_ARCH}.img ./
+    fi
     #shellcheck disable=SC2154
     archboot-cpio.sh -c "${MKINITCPIO_CONFIG}" -k "${ALL_kver}" -g "${_ISODIR}/boot/initramfs-${_ARCH}.img" || exit 1
     # delete cachedir on archboot environment
